+++ examples/react/package.json
@@ -15,18 +15,19 @@
     "@0xsequence/connect": "workspace:*",
     "@0xsequence/design-system": "2.0.12",
     "@0xsequence/network": ">= 2.2.13",
-    "@0xsequence/react-hooks": "workspace:*",
     "@0xsequence/wallet-widget": "workspace:*",
     "@0xsequence/immutable-connector": "workspace:*",
+    "@0xsequence/react-hooks": "workspace:*",
     "@0xsequence/waas": ">= 2.2.13",
-    "@tanstack/react-query": "^5.62.0",
     "@imtbl/config": "^2.1.2",
     "@imtbl/sdk": "^2.1.2",
+    "@tanstack/react-query": "^5.62.0",
     "clsx": "^2.1.1",
     "example-shared-components": "workspace:*",
     "motion": "^12.3.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.3.0",
     "tailwindcss": "^4.0.6",
     "typescript": "^5.8.2",
     "viem": "^2.23.10 ",
