+++ examples/react/src/App.tsx
@@ -1,8 +1,8 @@
 import { SequenceCheckoutProvider } from '@0xsequence/checkout'
 import { SequenceConnect } from '@0xsequence/connect'
 import { ThemeProvider } from '@0xsequence/design-system'
-import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import { SequenceWalletProvider } from '@0xsequence/wallet-widget'
+import { BrowserRouter, Routes, Route } from 'react-router-dom'
 
 import { Homepage } from './components/Homepage'
 import { config, checkoutConfig } from './config'
