+++ examples/react/src/config.ts
@@ -2,9 +2,8 @@ import { SequenceCheckoutConfig } from '@0xsequence/checkout'
 import { ConnectConfig, createConfig, WalletType } from '@0xsequence/connect'
 import { immutable } from '@0xsequence/immutable-connector'
 import { ChainId } from '@0xsequence/network'
-import { passport } from '@imtbl/sdk'
 import { Environment } from '@imtbl/config'
-
+import { passport } from '@imtbl/sdk'
 import { zeroAddress } from 'viem'
 
 const searchParams = new URLSearchParams(location.search)
