+++ packages/immutable-connector/src/wagmiConnector/index.ts
@@ -1,5 +1,5 @@
-import { passport } from '@imtbl/sdk'
 import { Environment } from '@imtbl/config'
+import { passport } from '@imtbl/sdk'
 import { type IMXProvider } from '@imtbl/x-provider'
 import { type Address } from 'viem'
 import { createConnector } from 'wagmi'
